requirements.txt
# Core dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0
google-generativeai>=0.6.0
python-dotenv>=1.0.0
streamlit>=1.23.0

# Document processing for API key extraction
python-docx>=0.8.11

# Testing
pytest>=7,<8
responses>=0.23.0

# Logging and error handling
logging>=0.5.1.2

# Type hints
typing-extensions>=4.5.0
